+++ src/control/getPost.js
@@ -1,13 +1,16 @@
 var fs = require('fs');
 
-var getPost = (callback) =>{
-	fs.readFile('./data/post.json', 'utf8', (err, data) => {
-		if (err) {
-			console.log("File read failed:", err);
-			return;
-		}
-		return JSON.parse(data);
-	});
+var getPost = (req,res) =>{
+	fs.readFile("./data/post.json", "utf8", (err, data) => {
+        if (err) {
+          console.log("File read failed:", err);
+          res.send("error");
+          return;
+        }
+		var params = req.params;
+        var ret = JSON.parse(data)[params.address][params.id]
+        res.status(200).json(ret);
+      });
 }
 
 module.exports = getPost;
